# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Novel Dialogue Reader API",
    "description": "轻小说对话辅助阅读器本地后端。契约摘要见 docs/CONTRACTS.md，完整规格见 DEVELOPMENT.md。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/books": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "书籍列表",
        "operationId": "list_books_route_api_books_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 200,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_CursorPage_BookOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{book_id}": {
      "get": {
        "tags": [
          "books"
        ],
        "summary": "书籍详情",
        "operationId": "get_book_route_api_books__book_id__get",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_BookOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/books/{book_id}/estimates": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "本地估算（不调用模型）",
        "operationId": "estimate_route_api_books__book_id__estimates_post",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_EstimateOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{book_id}/usage": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "按任务汇总用量",
        "operationId": "usage_route_api_books__book_id__usage_get",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_UsageOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/books/{book_id}/exports/preview": {
      "post": {
        "tags": [
          "exports"
        ],
        "summary": "冻结导出快照并返回样张（不调用模型）",
        "operationId": "export_preview_route_api_books__book_id__exports_preview_post",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportPreviewIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_ExportPreviewOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{book_id}/exports": {
      "post": {
        "tags": [
          "exports"
        ],
        "summary": "按快照生成 EPUB/HTML（幂等，本地执行）",
        "operationId": "create_export_route_api_books__book_id__exports_post",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Book Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportCreateIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_ExportArtifactOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exports/{export_id}": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "导出状态、校验结果与下载可用性",
        "operationId": "get_export_route_api_exports__export_id__get",
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_ExportArtifactOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exports/{export_id}/download": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "受控下载（从不覆盖原书）",
        "operationId": "download_export_route_api_exports__export_id__download_get",
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BookFormat": {
        "type": "string",
        "enum": [
          "TXT",
          "EPUB"
        ],
        "title": "BookFormat"
      },
      "BookOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "format": {
            "$ref": "#/components/schemas/BookFormat"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "import_status": {
            "$ref": "#/components/schemas/ImportStatus"
          },
          "read_position_cp": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Read Position Cp"
          },
          "reading_mode": {
            "$ref": "#/components/schemas/ReadingMode",
            "default": "initial"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "active_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Version Id"
          },
          "active_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookVersionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "format",
          "source_sha256",
          "import_status",
          "read_position_cp",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "BookOut"
      },
...
      "BudgetIn": {
        "properties": {
          "max_input_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Input Tokens"
          },
          "max_output_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Output Tokens"
          },
          "max_rechecks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Rechecks",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BudgetIn"
      },
...
      "CursorPage_BookOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BookOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CursorPage[BookOut]"
      },
...
      "DataEnvelope_BookOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BookOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[BookOut]"
      },
...
      "DataEnvelope_CursorPage_BookOut__": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CursorPage_BookOut_"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[CursorPage[BookOut]]"
      },
...
      "DataEnvelope_EstimateOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EstimateOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[EstimateOut]"
      },
      "DataEnvelope_ExportArtifactOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExportArtifactOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[ExportArtifactOut]"
      },
      "DataEnvelope_ExportPreviewOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExportPreviewOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[ExportPreviewOut]"
      },
...
      "DataEnvelope_UsageOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UsageOut"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "本次请求的追踪 ID，与响应头 X-Request-ID 一致。"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "request_id"
        ],
        "title": "DataEnvelope[UsageOut]"
      },
...
      "EstimateIn": {
        "properties": {
          "book_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Version Id"
          },
          "range": {
            "additionalProperties": true,
            "type": "object",
            "title": "Range"
          },
          "reading_mode": {
            "$ref": "#/components/schemas/ReadingMode",
            "default": "initial"
          },
          "visible_horizon_cp": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Visible Horizon Cp"
          },
          "budget": {
            "$ref": "#/components/schemas/BudgetIn"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "EstimateIn"
      },
      "EstimateOut": {
        "properties": {
          "book_id": {
            "type": "string",
            "title": "Book Id"
          },
          "book_version_id": {
            "type": "string",
            "title": "Book Version Id"
          },
          "window_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Window Count"
          },
          "target_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Target Count"
          },
          "input_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Input Tokens"
          },
          "output_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Output Tokens"
          },
          "total_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Tokens"
          },
          "estimator": {
            "additionalProperties": true,
            "type": "object",
            "title": "Estimator"
          },
          "policy": {
            "additionalProperties": true,
            "type": "object",
            "title": "Policy"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "book_id",
          "book_version_id",
          "window_count",
          "target_count",
          "input_tokens",
          "output_tokens",
          "total_tokens"
        ],
        "title": "EstimateOut"
      },
...
      "ExportCreateIn": {
        "properties": {
          "snapshot_id": {
            "type": "string",
            "title": "Snapshot Id"
          },
          "format": {
            "$ref": "#/components/schemas/ExportFormat"
          },
          "style": {
            "$ref": "#/components/schemas/ExportStyleIn"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Idempotency Key"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "snapshot_id",
          "format",
          "idempotency_key"
        ],
        "title": "ExportCreateIn"
      },
      "ExportFormat": {
        "type": "string",
        "enum": [
          "epub",
          "html"
        ],
        "title": "ExportFormat"
      },
      "ExportPreviewIn": {
        "properties": {
          "book_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Version Id"
          },
          "chapter_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Ids",
            "description": "留空表示整本；否则只导出选中的章节"
          },
          "visibility_policy": {
            "$ref": "#/components/schemas/VisibilityPolicy",
            "default": "position_safe"
          },
          "style": {
            "$ref": "#/components/schemas/ExportStyleIn"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ExportPreviewIn",
        "description": "冻结快照 + 样张；**不调用模型**。"
      },
...
      "ExportStyleIn": {
        "properties": {
          "preset": {
            "$ref": "#/components/schemas/ExportStylePreset",
            "description": "color_and_label / color_only / label_only",
            "default": "color_and_label"
          },
          "palette_id": {
            "type": "string",
            "maxLength": 64,
            "title": "Palette Id",
            "default": "reader-default"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ExportStyleIn"
      },
      "ExportStylePreset": {
        "type": "string",
        "enum": [
          "color_and_label",
          "color_only",
          "label_only"
        ],
        "title": "ExportStylePreset"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReadingMode": {
        "type": "string",
        "enum": [
          "initial",
          "reread"
        ],
        "title": "ReadingMode"
      },
...
      "UsageOut": {
        "properties": {
          "book_id": {
            "type": "string",
            "title": "Book Id"
          },
          "runs": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Runs"
          },
          "unknown_usage_runs": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unknown Usage Runs"
          },
          "input_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Input Tokens"
          },
          "output_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Output Tokens"
          },
          "total_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Tokens"
          },
          "by_state": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By State"
          },
          "by_model": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Model"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "book_id",
          "runs",
          "unknown_usage_runs",
          "input_tokens",
          "output_tokens",
          "total_tokens"
        ],
        "title": "UsageOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VisibilityPolicy": {
        "type": "string",
        "enum": [
          "position_safe",
          "reread"
        ],
        "title": "VisibilityPolicy"
      },
...
    }
  }
}